stock/polygon/DockerBuild/lambda/polystocks/requirements.txt
polygon-api-client
python-dotenv
streamlit
datetime
numerize
numpy
pandas
awslambdaric
